Added files via upload
The fact is, you never know when disaster is going to strike, be it illness, and accident, a crash in the stock market or the Vancouver mega earthquake. Yet so many of us live in constant anxiety of whats around the corner and it prevents us from enjoying the present. My device is there to do the worrying and waiting for you, so you can seize the day while it awaits danger. Beyond that is serves as a message, a cautionary tale if you will, of the futility of laying dormant in case of emergency. Through the use of vibration sensors, an Arduino and a sleek simple speaker, the Quaketrack sits calmly on your mantle; until it doesn’t. The sensors will trigger the Arduino if tremors over a 5.5 are felt, queuing the speaker to blast the soundtrack of your triumph over adversity. Wether you’re inclined to fight or flight, this speaker will accompany you through both, sure to alert you to the presence of the earthquake day or night, rain or shine. Upon purchasing and charging up your Quaket